fix: resolved eslint errors
